backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
langgraph>=0.0.65
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-anthropic>=0.1.15
langchain-google-genai>=1.0.3
google-genai>=0.2.2
python-dotenv>=1.0.1
pandas>=2.2.0
pydantic>=2.7.0
python-multipart>=0.0.9
email-validator>=2.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openpyxl>=3.1.0
